@@ -16,11 +16,11 @@
   "devDependencies": {
     "@sveltejs/adapter-auto": "3.2.0",
     "@sveltejs/kit": "2.8.3",
-    "@sveltejs/vite-plugin-svelte": "^3.1.0",
-    "svelte": "4.2.19",
-    "svelte-check": "3.6.9",
-    "typescript": "5.4.5",
-    "vite": "5.2.14"
+    "@sveltejs/vite-plugin-svelte": "^4.0.0",
+    "svelte": "^5.0.0",
+    "svelte-check": "^4.0.0",
+    "typescript": "^5.5.0",
+    "vite": "^5.4.4"
   },
   "dependencies": {
     "@auth/sveltekit": "latest"
